+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Java Bag Reader changelog
 
+1.2
+
+- Removing hard-coded GPS and vehicle name topics
+- Adding a better API for iterating through message types and topics
+
 1.1
 
 - Adding examples to the README